Split watch.rs into watch/{mod,stack_trace}.rs and automatically run flamegraph.pl
