+++ .gitignore
@@ -1,2 +1,2 @@
 /target
-flame_graph.txt
+flame_graph*
